+++ altair/vega/v3/__init__.py
@@ -1,2 +1,8 @@
 from .display import vega, Vega, renderers
 from .schema import *
+
+from .data import (
+    pipe, curry, limit_rows,
+    sample, to_json, to_csv, to_values,
+    default_data_transformer
+)
