+++ altair/vegalite/v1/data.py
@@ -16,3 +16,17 @@ data_transformers.register('default', default_data_transformer)
 data_transformers.register('json', to_json)
 data_transformers.register('csv', to_csv)
 data_transformers.enable('default')
+
+
+__all__ = (
+    'MaxRowsError',
+    'curry',
+    'default_data_transformer',
+    'limit_rows',
+    'pipe',
+    'sample',
+    'to_csv',
+    'to_json',
+    'to_values',
+    'data_transformers'
+)
